% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{streamxform}
\alias{streamxform}
\title{transform 3D points using one or more CMTK registrations}
\usage{
streamxform(points, reglist, inversionTolerance = 1e-08,
  affineonly = FALSE)
}
\arguments{
\item{points}{an Nx3 matrix of 3D points}

\item{reglist}{A character vector specifying registrations. See details.}

\item{inversionTolerance}{the precision of the numerical inversion when
transforming in the inverse direction.}

\item{affineonly}{Whether to apply only the affine portion of transforms
default \code{FALSE}.}
}
\description{
transform 3D points using one or more CMTK registrations
}
\details{
To transform points from sample to reference space, you will need
  to use the inverse transformation. This can be achieved by preceding the
  registration with a \verb{--inverse} flag. When multiple registrations are
  being used the are ordered from sample to reference brain.
}
\examples{
\dontrun{
m=matrix(rnorm(30,mean = 50), ncol=3)
reg=system.file("extdata","cmtk","FCWB_JFRC2_01_warp_level-01.list", package='cmtkr')
# from reference to sample
streamxform(m, reg)

# from sample to reference
streamxform(m, c("--inverse", reg))

# concatenating 3 registrations to map S -> B1 -> B2 -> T
# the first two registrations are inverted, the last is not.
streamxform(m, c("--inverse", StoB1, "--inverse", B1toB2, TtoB2))
}
}
