% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtkr-package.r
\docType{package}
\name{cmtkr-package}
\alias{cmtkr-package}
\alias{cmtkr}
\title{R wrapper for CMTK}
\description{
Provides R bindings to selected components of the Computational Morphometry Toolkit ('CMTK') for image registration and point transformation. A subset of the C++ source code required for point transforms is bundled with 'cmtkr'. This allows direct calls into the 'CMTK' library, avoiding command-line invocations and providing order-of-magnitude speed improvements. Additional 'CMTK' functionality may be wrapped in future releases. 'CMTK' is described in Rohlfing T and Maurer CR (2003) \doi{10.1109/titb.2003.808506}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/cmtkr}
  \item Report bugs at \url{https://github.com/jefferis/cmtkr/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

Authors:
\itemize{
  \item Torsten Rohlfing (CMTK library author (included source))
}

Other contributors:
\itemize{
  \item SRI International (Copyright holder for CMTK library) [copyright holder]
  \item Google, Inc. (Copyright holder for CMTK library contributions) [copyright holder]
}

}
\keyword{package}
